backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlmodel==0.0.24
python-jose[cryptography]==3.4.0
cryptography==44.0.3
httpx==0.28.1
otf-api>=0.15.0
stravalib>=2.0
fit-tool>=0.9.13
python-multipart==0.0.20
pydantic-settings==2.9.1
alembic==1.15.2
aiosqlite==0.21.0
